requirements.txt
matplotlib==3.5.3
matplotlib==3.5.2
numpy==1.21.5
opencv_python==4.7.0.72
pandas==1.3.5
pandas==1.4.4
Pillow==9.4.0
Pillow==9.2.0
Pillow==10.0.0
tensorflow==2.8.1
